+++ beacon-chain/rpc/debug/state.go
@@ -3,7 +3,6 @@ package debug
 import (
 	"context"
 
-	ptypes "github.com/gogo/protobuf/types"
 	"github.com/prysmaticlabs/go-ssz"
 	pbrpc "github.com/prysmaticlabs/prysm/proto/beacon/rpc/v1"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
@@ -61,14 +60,4 @@ func (ds *Server) GetBeaconState(
 	default:
 		return nil, status.Error(codes.InvalidArgument, "Need to specify either a block root or slot to request state")
 	}
-}
-
-// GetProtoArrayForkChoice returns proto array fork choice store.
-func (ds *Server) GetProtoArrayForkChoice(ctx context.Context, _ *ptypes.Empty) (*pbrpc.ProtoArrayForkChoiceResponse, error) {
-	store := ds.HeadFetcher.ProtoArrayStore()
-	return &pbrpc.ProtoArrayForkChoiceResponse{
-		PruneThreshold: store.PruneThreshold,
-		JustifiedEpoch: store.JustifiedEpoch,
-		FinalizedEpoch: store.FinalizedEpoch,
-	}, nil
 }
